requirements.txt
openai
pandas
numpy
folium
opencv-python
base64
streamlit_folium
streamlit_chat